Move cleanTags to function
